:dizzy: Move around several files
- `Rem.Wordle` -> `Wordle`
Wordle seems more like a mini stand_alone domain that can play a game.
It still depends on Rem, but I think Rem should not be concerned with
how the game is played. (Though I feel knowing the coordination is ok)

- `Rem.Models` -> `Schema`
I just like the Schema name more. I'm still unsure if I should have left
it scoped in Rem, but I feel it's a bit cleaner to have it like this for
now...

This one was faster to do than I expected, I still have to do some
refactoring around the commands and the sessions, but it probably is
easier now to figure out how to design it.