:dizzy: Refactored commands
Moved things around a bit more.
I'm happy with the general design, but I feel the naming is still a bit
confusing.

This is the basic idea:
Consumer -> PrefixCommand/SessionHandler <-> AnythingElse

My one complaint with the current layout of things is that it's not
super obvious which files are prefix commands and which ones are session
handlers.

I feel I can rename the concept `command` into something more generic so
I can separate the other types better. I still think they need to be
together in the same folder because they do share some functionality.
And their difference in design is really not that big.

I'll figure something out, but for the time being I'll write tests and
fix some other stuff.

I apologize in advance if I end up doing another rather large refactor.