:sparkles: Created a cronjob that fetches solutions
Once a day and also on boot seemed like a really good setup
Fetches solutions that are missing in the DB, based on the last saved
number.

Since we can only fetch them sequentially I batch them rudimentarily. We
make 10 hits and then sleep for a second. Yeah... not the best but we
will be normally fetching one per day.